refactor(capi): extract CAPI_internal.h (session ctx + boundary guard), kill try/catch boilerplate
Behaviour-preserving first step of the todo.md Phase 2 file split:

- struct dasher_ctx, small utils (nowMs/lround_int/clamp_int/colorToARGB),
  inputTime moved to src/CAPI_internal.h (moved, not copied; internal-only,
  not installed, zero ABI impact). Interface defined out-of-line as
  dasher_ctx::Interface in CAPI.cpp.
- capi::guarded / capi::guarded_result templates replace the 15 uniform
  try/catch sites (~30 duplicated catch blocks). Same log text, same
  engineError latching. Sites with bespoke or deliberately-silent handling
  keep explicit catches (listed in CAPI.cpp header comment).
- diagnostic delta: reset_settings log messages gain a context prefix.

Verified: full rebuild, 45/45 ctest, symbol freeze 109 unchanged.
